fixed parseInt in  math.js to parseFloat

+++ utils/mathUtilities.js
@@ -1,8 +1,3 @@
-/**
- * Gets a random multiplication, division, subtraction or addition question
- *
- * @returns {} The randomly generated math question
- */
 function getQuestion() {
   const operations = ['+', '-', '*', '/'];
   const randomOperations =
@@ -30,29 +25,22 @@ function getQuestion() {
   return { question, answer };
 }
 
-/**
- * Parses the provided question and gets whether or not the provided answer is correct
- *
- * @param {*} question The question being answered
- * @param {*} answer The potential answer
- * @returns {boolean} True if the answer was correct, false otherwise.
- */
 function isCorrectAnswer(question, answer) {
   const [num1, operator, num2] = question.split(' ');
   let correctAnswer;
 
   switch (operator) {
     case '+':
-      correctAnswer = parseInt(num1) + parseInt(num2);
+      correctAnswer = parseFloat(num1) + parseFloat(num2);
       break;
     case '-':
-      correctAnswer = parseInt(num1) - parseInt(num2);
+      correctAnswer = parseFloat(num1) - parseFloat(num2);
       break;
     case '*':
-      correctAnswer = parseInt(num1) * parseInt(num2);
+      correctAnswer = parseFloat(num1) * parseFloat(num2);
       break;
     case '/':
-      correctAnswer = parseInt(num1) / parseInt(num2);
+      correctAnswer = parseFloat(num1) / parseFloat(num2);
       break;
   }
